# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fetch — AI-powered food discovery and nutrition management
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /scans/analyze:
    post:
      tags: [Scans]
      operationId: analyzeScan
      summary: Analyze a food photo with vision AI and persist the result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanInput'
      responses:
        '200':
          description: Analysis complete.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: AI provider error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /scans:
    get:
      tags: [Scans]
      operationId: listScans
      summary: List the user's scan history
      responses:
        '200':
          description: Scans loaded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scan'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /scans/{id}:
    get:
      tags: [Scans]
      operationId: getScan
      summary: Get a single scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scan loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Scans]
      operationId: deleteScan
      summary: Delete a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted.
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: integer, minimum: 1 }
        contentType: { type: string, minLength: 1 }
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string, format: uri }
        objectPath: { type: string }
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...
    ScanInput:
      type: object
      required: [imageObjectPath]
      properties:
        imageObjectPath:
          type: string
          description: Object path returned from `POST /storage/uploads/request-url` after the client PUTs the photo to GCS.
          minLength: 1
        note:
          type: string
          description: Optional user note to bias analysis (e.g. "lunch portion").
    ScanIngredient:
      type: object
      required: [name, amount]
      properties:
        name: { type: string }
        amount: { type: string }
    DietaryCompliance:
      type: object
      required: [vegetarian, vegan, glutenFree, dairyFree, pescatarian, keto]
      properties:
        vegetarian: { type: boolean }
        vegan: { type: boolean }
        glutenFree: { type: boolean }
        dairyFree: { type: boolean }
        pescatarian: { type: boolean }
        keto: { type: boolean }
    AllergenWarning:
      type: object
      required: [allergen, severity, reason]
      properties:
        allergen: { type: string }
        severity:
          type: string
          enum: [trace, contains, may_contain]
        reason: { type: string }
    Scan:
      type: object
      required:
        - id
        - userId
        - imageObjectPath
        - foodName
        - description
        - calories
        - protein
        - carbs
        - fat
        - fiber
        - sugar
        - healthScore
        - environmentalScore
        - dietaryCompliance
        - allergens
        - ingredients
        - tags
        - createdAt
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        imageObjectPath:
          type: string
          description: Object storage path of the uploaded photo, e.g. `/objects/uploads/uuid`. Serve via `GET /api/storage{imageObjectPath}`.
        foodName: { type: string }
        description: { type: string }
        calories: { type: integer }
        protein: { type: number }
        carbs: { type: number }
        fat: { type: number }
        fiber: { type: number }
        sugar: { type: number }
        healthScore:
          type: integer
          minimum: 0
          maximum: 100
        environmentalScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Environmental impact score (0-100, higher is better/lower-impact).
        dietaryCompliance:
          $ref: '#/components/schemas/DietaryCompliance'
        allergens:
          type: array
          items:
            $ref: '#/components/schemas/AllergenWarning'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/ScanIngredient'
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
...